review modifications: maniamodmirror inheritance, reflection utilities, vertical flip option
